Fixed intialization bug introduced when rewriting code to suit compiler
The newScopeGraph function contained a tricky circular dependency
between variables that are used in nested functions. Assessing this case
as highly exotic I decided to rewrite the code and break the cycle
instead of performing a heavy analysis in the compiler.

Alas, the rewrite missed the proper iitialization of these variables.